service_backend/requirements.txt
argon2-cffi==19.1.0
cffi==1.13.0
Click==7.0
dnspython==1.16.0
email-validator==1.0.4
fastapi==0.42.0
h11==0.8.1
idna==2.8
peewee==3.11.2
psycopg2==2.8.3
pycparser==2.19
pydantic==0.32.2
PyJWT==1.7.1
six==1.12.0
starlette==0.12.9
uvicorn==0.9.0
websockets==8.0.2
